Add files via upload
i don't know if it's wrong or not because i have some doubt about the fact that the graph generate are random (so the number of nodes changes everytime). in the file i uses the function created and i tried to plot the complexity (i use running time). BFS seems to be the most efficient.